hm/superion: fix audio function keys _again_
It turns out that volume up/down/mute-toggle on the easyeffects_sink
doesn't actually do anything. It succeeds, and the OSD works, but
it doesn't actually change volume levels. Oops.

This removes the shell script wrapper for swayosd-client and instead
overrides sway keybindings to pass `--device` with the real ALSA sink
name every time, which does work.

Or it would, if swayosd worked properly, so my silly patch is still
in place (replaced with a parameterized one which I `substituteAll`
in order to only specify the audio device name in one place).